server-code/python-flask/httpbin/openapi_server/controllers/dynamic_data_controller.py
import connexion
from typing import Dict
from typing import Tuple
from typing import Union

from openapi_server import util


def base64_value_get(value):  # noqa: E501
    """Decodes base64url-encoded string.

     # noqa: E501

    :param value: 
    :type value: str

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def bytes_nget(n):  # noqa: E501
    """Returns n random bytes generated with given seed

     # noqa: E501

    :param n: 
    :type n: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_delete(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_get(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_patch(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_post(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_put(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def delay_delay_trace(delay):  # noqa: E501
    """Returns a delayed response (max of 10 seconds).

     # noqa: E501

    :param delay: 
    :type delay: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def drip_get(duration=None, numbytes=None, code=None, delay=None):  # noqa: E501
    """Drips data over a duration after an optional initial delay.

     # noqa: E501

    :param duration: The amount of time (in seconds) over which to drip each byte
    :type duration: 
    :param numbytes: The number of bytes to respond with
    :type numbytes: int
    :param code: The response code that will be returned
    :type code: int
    :param delay: The amount of time (in seconds) to delay before responding
    :type delay: 

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def links_n_offset_get(n, offset):  # noqa: E501
    """Generate a page containing n links to other pages which do the same.

     # noqa: E501

    :param n: 
    :type n: int
    :param offset: 
    :type offset: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def range_numbytes_get(numbytes):  # noqa: E501
    """Streams n random bytes generated with given seed, at given chunk size per packet.

     # noqa: E501

    :param numbytes: 
    :type numbytes: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def stream_bytes_nget(n):  # noqa: E501
    """Streams n random bytes generated with given seed, at given chunk size per packet.

     # noqa: E501

    :param n: 
    :type n: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def stream_nget(n):  # noqa: E501
    """Stream n JSON responses

     # noqa: E501

    :param n: 
    :type n: int

    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'


def uuid_get():  # noqa: E501
    """Return a UUID4.

     # noqa: E501


    :rtype: Union[None, Tuple[None, int], Tuple[None, int, Dict[str, str]]
    """
    return 'do some magic!'
